Add type.string
