+++ pseudo/type/numbers.py
@@ -5,14 +5,6 @@ __author__ = "Patryk Niedźwiedziński"
 from pseudo.type.base import Value
 
 
-def is_digit(c) -> bool:
-    """Checks if given char is a digit."""
-    try:
-        return ord(c) >= 48 and ord(c) <= 57
-    except TypeError:
-        return False
-
-
 class Int(Value):
     """Int value node."""
 
@@ -24,3 +16,22 @@ class Int(Value):
 
     def __repr__(self):
         return f"Int({repr(self.value)})"
+
+
+def is_digit(c) -> bool:
+    """Checks if given char is a digit."""
+    try:
+        return ord(c) >= 48 and ord(c) <= 57
+    except TypeError:
+        return False
+
+
+def read_number(lexer) -> Int:
+    """Read a number from the stream."""
+    number = lexer.read(is_digit)
+    try:
+        int(number)
+    except ValueError:
+        return None
+    return Int(number)
+
