+++ tests/conftest.py
@@ -1,0 +1,9 @@
+import pytest
+import pseudo
+
+
+@pytest.fixture
+def lexer():
+    """Returns lexer object."""
+    lex = pseudo.lexer.Lexer("")
+    return lex
