+++ tests/type_test/numbers_test.py
@@ -2,7 +2,8 @@
 
 __author__ = "Patryk Niedźwiedziński"
 
-from pseudo.type.numbers import is_digit
+from pseudo.type.numbers import is_digit, read_number
+from pseudo.stream import Stream
 
 
 def test_is_digit():
@@ -16,3 +17,14 @@ def test_is_digit():
     if is_digit('"') is True:
         print(is_digit('"'))
         raise AssertionError
+
+
+def test_read_number(lexer):
+    """Checks read_number"""
+    lexer.i = Stream("123")
+    if 123 != read_number(lexer).value:
+        raise AssertionError
+    lexer.i = Stream("abc")
+    if read_number(lexer) is not None:
+        raise AssertionError
+
