+++ tests/type_test/string_test.py
@@ -1,0 +1,11 @@
+"""This module contains tests for pseudo.type.string"""
+
+from pseudo.type.string import read_string
+from pseudo.stream import Stream
+
+
+def test_read_string(lexer):
+    """Checks Lexer.read_string"""
+    lexer.i = Stream('"abc"')
+    if "abc" != read_string(lexer).value:
+        raise AssertionError
